+++ src/main/kotlin/com/wiggglez/kt_stockdata_lib/historicchartdata/dataobjects/yfhistoryresp/Quote.kt
@@ -13,5 +13,5 @@ internal data class Quote(
     @SerializedName("open")
     val `open`: List<Double?>,
     @SerializedName("volume")
-    val volume: List<Int?>
+    val volume: List<Long?>
 )
